#137 refactor: 불필요 코드 삭제, 타입 수정
