+++ src/page/archiving/index/hook/api/useGetTimeBlockQuery.ts
@@ -2,9 +2,9 @@ import { useQuery } from '@tanstack/react-query';
 
 import { getTimeline } from '@/shared/api/archiving/time-blocks';
 
-export const useGetTimeBlockQuery = (timeId: number, timeline: string, date: string) => {
+export const useGetTimeBlockQuery = (teamId: number, timeline: string, date: string) => {
   return useQuery({
-    queryKey: ['timeBlock', date],
-    queryFn: () => getTimeline(timeId, timeline, date),
+    queryKey: ['timeBlock', teamId, date],
+    queryFn: () => getTimeline(teamId, timeline, date),
   });
 };
