+++ src/page/showcase/index/ShowcasePage.tsx
@@ -2,6 +2,7 @@ import { chipStyle, clubProfileStyle, containerStyle } from '@/page/showcase/ind
 import BannerSlider from '@/page/showcase/index/component/BannerSlider/BannerSlider';
 import CategoryChip from '@/page/showcase/index/component/CategoryChip/CategoryChip';
 import ClubProfileCard from '@/page/showcase/index/component/ClubProfileCard/ClubProfileCard';
+import { useClubListQuery } from '@/page/showcase/index/hook/api/useClubListQuery';
 
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
@@ -9,8 +10,7 @@ import { useNavigate } from 'react-router-dom';
 import Flex from '@/common/component/Flex/Flex';
 import Heading from '@/common/component/Heading/Heading';
 
-import useCategoryListQuery from '@/shared/hook/useCategoryListQuery';
-import { useClubListQuery } from '@/shared/hook/useClubListQuery';
+import useCategoryListQuery from '@/shared/hook/api/useCategoryListQuery';
 
 const ShowcasePage = () => {
   const navigate = useNavigate();
